requirements.txt
googleapis-common-protos==1.65.0
grpclib==0.4.9
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
load-dotenv==0.1.0
multidict==6.1.0
numpy==2.1.2
opencv-python-headless==4.10.0.84
pillow==10.4.0
protobuf==5.28.2
python-dotenv==1.0.1
pyzbar==0.1.9
qrcode==8.0
typing_extensions==4.12.2
viam-sdk==0.31.0
